Maybe fix CI
